yolov8 refactored to Ultralytics
