+++ python/ultralytics_to_onnx.py
@@ -7,16 +7,16 @@ import os
 import sys
 
 
-MODEL_SIZE = sys.argv[1]
+MODEL_NAME = sys.argv[1]
 IMAGE_SIZE = 640
-PT_MODEL_URL = f"https://github.com/ultralytics/assets/releases/download/v8.3.0/yolov8{MODEL_SIZE}.pt"
+PT_MODEL_URL = f"https://github.com/ultralytics/assets/releases/download/v8.3.0/{MODEL_NAME}.pt"
 
 export_dir = "models"
 
 
-pt_file = f"{export_dir}/yolov8{MODEL_SIZE}.pt"
-model_file = f"{export_dir}/yolov8{MODEL_SIZE}.onnx"
-classes_file = f"{export_dir}/yolov8{MODEL_SIZE}_classes.json"
+pt_file = f"{export_dir}/{MODEL_NAME}.pt"
+model_file = f"{export_dir}/{MODEL_NAME}.onnx"
+classes_file = f"{export_dir}/{MODEL_NAME}_classes.json"
 
 
 # making `export_dir`
@@ -26,7 +26,7 @@ except OSError as error:
     print(export_dir + ": directory already created")
 
 def download_pt_model(url, path):
-  print(f"Downloading YoloV8{MODEL_SIZE} .pt file")
+  print(f"Downloading {MODEL_NAME}.pt file")
   with requests.get(url, stream=True) as r:
     r.raise_for_status()
     with open(path, 'wb') as fp:
@@ -42,7 +42,7 @@ model_path = download_pt_model(PT_MODEL_URL, pt_file)
 model = YOLO(pt_file)
 
 # Export the models
-model.export(format='onnx', imgsz=IMAGE_SIZE, opset=9)
+model.export(format='onnx', imgsz=IMAGE_SIZE, opset=12)
 
 # Export the categories
 with open(classes_file, "w") as f:
